Different ranges according to datasource
Each head node type has its own amount of information to send to the distributed workers. For example, if the RDataFrame is just building on some empty entries, it is enough to send Range objects made of a pair of integers (start, end). When a TTree is involved, it has at least a name and a file but might also have information about friend trees.
These differences are now reflected in different types of Range objects, i.e. EntriesRange and TreeRange for now. Furthermore, there is no need to host the logic to create Ranges in a class. So it is now hosted on its own Python module with free functions. This makes it more modular and would possibly also allow for caching the Range objects for subsequent usages.
